backend/clustering/requirements.txt
tensorflow==2.4.1
pandas==1.2.3
numpy==1.20.1
scipy==1.4.1
attrs==19.3.0